Move UseLayout into utils/
Also move useUpdateEffect.
